added schema in settings
