+++ src/main/java/com/online/store/dao/impl/ProductDaoImpl.java
@@ -2,7 +2,6 @@ package com.online.store.dao.impl;
 
 import com.online.store.dao.ProductDao;
 import com.online.store.db.Storage;
-import com.online.store.lib.Dao;
 import com.online.store.model.Product;
 import java.util.List;
 import java.util.Optional;
