+++ src/main/java/com/online/store/exceptions/DataProcessingException.java
@@ -1,0 +1,7 @@
+package com.online.store.exceptions;
+
+public class DataProcessingException extends RuntimeException {
+    public DataProcessingException(String message, Throwable e) {
+        super(message, e);
+    }
+}
