+++ src/main/java/com/online/store/util/ConnectionUtil.java
@@ -18,14 +18,11 @@ public class ConnectionUtil {
         Properties dbProperties = new Properties();
         dbProperties.put("user", "root");
         dbProperties.put("password", "1111");
-        //String url = "jdbc:mysql://localhost:3306/online_store";
         String url = "jdbc:mysql://localhost:3306/online_store?serverTimezone=UTC";
         try {
-            Connection connection = DriverManager.getConnection(url, dbProperties);
-            System.out.println("Connection to DB established");
-            return connection;
+            return DriverManager.getConnection(url, dbProperties);
         } catch (SQLException e) {
-            throw new RuntimeException("Can't established the connection to DB" , e);
+            throw new RuntimeException("Can't established the connection to DB", e);
         }
     }
 }
